+++ src/components/shared/index.ts
@@ -3,6 +3,12 @@ export * from './separator';
 export * from './search-input';
 export * from './guide';
 export * from './pagination';
+export * from './popover-backdrop';
+export * from './script-with-cleanup';
+export * from './video-card';
+export * from './video-list';
+export * from './video-thumbnail';
+export * from './with-auth';
 
 // eslint no-cycle 오류 때문에 export 안함
 // export * from './layout-with-header';
