+++ src/components/watch-page/index.ts
@@ -1,4 +1,3 @@
 export * from './video-info-section';
-export * from './popover-backdrop';
 export * from './user-menu';
 export * from './watch-page-header';
